pyproject.toml
[project]
name = "isp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-phonenumber-field>=8.4.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "phonenumbers>=9.0.21",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.2.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "routeros-api>=0.21.0",
    "uvicorn>=0.40.0",
    "whitenoise>=6.9.0",
]
